TF-1427 Move composer button to dashboard view on web
(cherry picked from commit 4f8da9cb313a0aa0f2a881de109ed64d3d284358)
